WIP ARQ-1957 Check for any caught Exception on client side
When running in Standalone mode the returned TestResult exception
does not go through the remote JUnitRunner which in this case clean
up some of the exception cases. We need to handle more then just
AfterJUnitExceptions.

Note: Needs more testing to verify we're not returning faulty exceptions when
e.g. @Test(expected=Exception) or @Rule Excepected is used.